+++ pages/index.tsx
@@ -2,11 +2,11 @@ import Head from "next/head";
 import Random from "../components/Random";
 import Footer from "../components/Footer";
 import Header from "../components/Header";
-import ArtistList from "../components/ArtistList";
-import FavList from "../components/FavList";
+import ArtistView from "../components/ArtistView";
 import Button from "../components/Button";
 import { useState } from "react";
 import styled from "styled-components";
+import FavView from "../components/FavView";
 
 type HomeProps = {
   onLike: () => void;
@@ -17,10 +17,11 @@ export default function Home({ onLike, likes }: HomeProps): JSX.Element {
   //Typescript
   type ViewPoint = { random: boolean; artists: boolean; favorites: boolean };
   //Typescipt end
+
   const [viewPoint, setViewPoint] = useState<ViewPoint>({
     random: true, //the standard view. It's called "random" because later I will implement a function to show a random collection of tattoos
     artists: false,
-    favorites: false, //I don't need this right now, but in the future
+    favorites: false,
   });
 
   function handleSwitchView(view: string): void {
@@ -54,8 +55,8 @@ export default function Home({ onLike, likes }: HomeProps): JSX.Element {
         />
       </StyledButtonWrapper>
       {viewPoint.random && <Random />}
-      {viewPoint.artists && <ArtistList onLike={onLike} likes={likes} />}
-      {viewPoint.favorites && <FavList onLike={onLike} likes={likes} />}
+      {viewPoint.artists && <ArtistView onLike={onLike} likes={likes} />}
+      {viewPoint.favorites && <FavView onLike={onLike} likes={likes} />}
       <Footer />
     </>
   );
